requirements.txt
Django==1.10.1
django-appconf==1.0.2
django-bootstrap-form==3.2.1
git+https://github.com/pbucher/django-bootstrap-datepicker.git@b6ebbd8bf557066e1
django-geoposition==0.3.0
django-localflavor==1.3
django-model-utils==2.6
django-user-accounts==1.3.1
django-formtools==1.0
Pillow==3.3.1
pinax-theme-bootstrap==7.10.2
pytz==2016.6.1

